pyproject.toml
[project]
name = "sqlalchemy-sandbox"
version = "0.1.0"
description = "testing sqlqlchemy options"
readme = "README.md"
authors = [{ name = "pr0f37 🚀", email = "adam.nowik@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.1",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg[c]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
]

[project.scripts]
sqlalchemy-sandbox = "sqlalchemy_sandbox:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
